wip: configurable pid params
